v0.8 Sprint 3 (QA-016): decompose NativeTcpStream — extract 3 pkg-private seams
NativeTcpStream remains the central runtime stream but offloads three
self-contained concerns to pkg-private siblings in the same package:

- NativeTcpStreamPlainSocketIo (173 LOC) — owns Backend enum
  (core-socket-seam vs nio-fallback selection) and the Panama
  recv()/send() syscall dispatch with NIO fallback. Engine retains
  thin readPlainIngress / tryWrite wrappers that delegate by backend.
- NativeTcpStreamConsumerGate (69 LOC) — owns the single-consumer
  invariant helpers (acquire / tryAcquire / release) previously inlined
  as static methods on the engine.
- NativeTcpStreamPendingWrite (132 LOC) — owns the outbound plain +
  lazy ciphertext buffers and offsets across partial socket writes.
  Decoupled from the engine via a TryWriter functional interface and
  a TlsContext record (tlsEngine, tlsLock, allocator), so the engine's
  socket-backend selection stays encapsulated in PlainSocketIo.

Engine LOC: 1269 → 1101 (-13%). Class-level @SuppressWarnings: 8 → 7.
Closed: AvoidDeeplyNestedIfStmts (advanceInbound flattened to
adoptNextInboundFromQueue helper + named state constants
INBOUND_ADVANCE_*).

Residual: GodClass + TooManyMethods retained with rationale —
WMC=241 indicates further decomposition is needed (TLS handshake
state machine + inbound state machine are next candidates) but
those splits would create circular dependencies with engine
lifecycle and are deferred to QA-016b.

Public surface unchanged: TransportStream contract + carrier-facing
pkg-private callbacks (offerIngress / markRemoteClosed / signalWriteReady /
markRegistrationPending / markRegistrationReady / readTlsIngressFromFd /
decryptIngress / flushPendingWrites / readPlainIngress) all retained.

Verified:
- Full reactor verify -Pcoverage SUCCESS (1:52)
- Transport TCK suite 23/23 green (CarrierPinning, ZeroAlloc,
  MultiReactor2, MultiReactor4 variants)
- NativeTcpStream*Test 7/7 (CloseWakeup, IngressArrivalWakeup)
- PMD + Checkstyle 0 violations
- JaCoCo BUNDLE LINE coverage check met

Sprint 3 GodClass tracker after this PR:
- QA-015 CommunityHttpClientEngine ✅
- QA-016 NativeTcpStream (this PR) ✅
- QA-016b NativeTcpStream (handshake/inbound extraction) — optional
- QA-017 JdbcFlowSnapshotStore — next
- QA-018 CommunityHttp2SessionProcessor + SubsystemOrchestrator
